+++ src/mako/chrono/TimeZone.php
@@ -39,7 +39,7 @@ class TimeZone extends DateTimeZone
 
 	/**
 	 * Returns an array of grouped time zones where the key is
-	 * a valid PHP timezone while the value is a presentable name.
+	 * a valid PHP time zone while the value is a presentable name.
 	 *
 	 * @return array
 	 */